New chages were added. Users and Weapons models, dtos, services and controllers. Even relationships.

+++ Services/CharacterService/CharacterService.cs
@@ -3,6 +3,7 @@ global using AutoMapper;
 using System;
 using System.Collections.Generic;
 using System.Linq;
+using System.Security.Claims;
 using System.Threading.Tasks;
 
 namespace dotnet.Services.CharacterService
@@ -11,61 +12,74 @@ namespace dotnet.Services.CharacterService
     {
         private readonly IMapper _mapper;
         private readonly DataContext _context;
+        private readonly IHttpContextAccessor _httpContextAccessor;
 
-        public CharacterService(IMapper mapper, DataContext context){
+        public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
+        {
+            _httpContextAccessor = httpContextAccessor;
             _mapper = mapper;
             _context = context;
         }
 
-        public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(AddCharacterDto character)
+        private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext!.User
+            .FindFirstValue(ClaimTypes.NameIdentifier)!);
+
+        public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(AddCharacterDto newCharacter)
         {
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
-            var newCharacter =   _mapper.Map<Character>(character);
-            
-            _context.Characters.Add(newCharacter);       
+            var character = _mapper.Map<Character>(newCharacter);
+            character.User = await _context.Users.FirstOrDefaultAsync(u => u.Id == GetUserId());
+
+            _context.Characters.Add(character);
             await _context.SaveChangesAsync();
 
-            serviceResponse.Data = await _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
+            serviceResponse.Data =
+                await _context.Characters
+                    .Where(c => c.User!.Id == GetUserId())
+                    .Select(c => _mapper.Map<GetCharacterDto>(c))
+                    .ToListAsync();
             return serviceResponse;
         }
 
         public async Task<ServiceResponse<GetCharacterDto>> UpdateCharacter(UpdateCharacterDto updatedCharacter)
         {
             var serviceResponse = new ServiceResponse<GetCharacterDto>();
-            try{
-           var character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
-           if(character is null) throw new Exception($"Character with ID '{updatedCharacter.Id}' not found.");
+            try
+            {
+                var character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
+                if (character is null) throw new Exception($"Character with ID '{updatedCharacter.Id}' not found.");
 
-           character.Name = updatedCharacter.Name;
-           character.HitPoints = updatedCharacter.HitPoints;
-           character.Strength = updatedCharacter.Strength;
-           character.Defense = updatedCharacter.Defense;
-           character.Intelligence = updatedCharacter.Intelligence;
-           character.Class = updatedCharacter.Class;
+                character.Name = updatedCharacter.Name;
+                character.HitPoints = updatedCharacter.HitPoints;
+                character.Strength = updatedCharacter.Strength;
+                character.Defense = updatedCharacter.Defense;
+                character.Intelligence = updatedCharacter.Intelligence;
+                character.Class = updatedCharacter.Class;
 
-            await _context.SaveChangesAsync();  
-           serviceResponse.Data = _mapper.Map<GetCharacterDto>(character);
-            }  
-            catch(Exception ex){
+                await _context.SaveChangesAsync();
+                serviceResponse.Data = _mapper.Map<GetCharacterDto>(character);
+            }
+            catch (Exception ex)
+            {
                 serviceResponse.Success = false;
                 serviceResponse.Message = ex.Message;
             }
-           
-           return serviceResponse;
+
+            return serviceResponse;
         }
 
         public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
         {
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
-            var dbCharacters = await _context.Characters.ToListAsync();
-            serviceResponse.Data = dbCharacters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();;
+            var dbCharacters = await _context.Characters.Where(c => c.User!.Id == GetUserId()).ToListAsync();
+            serviceResponse.Data = dbCharacters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList(); ;
             return serviceResponse;
         }
 
         public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
         {
             var serviceResponse = new ServiceResponse<GetCharacterDto>();
-            var dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
+            var dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.User!.Id == GetUserId());
             serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter);
             return serviceResponse;
         }
@@ -73,21 +87,23 @@ namespace dotnet.Services.CharacterService
         public async Task<ServiceResponse<List<GetCharacterDto>>> DeleteCharacter(int id)
         {
             var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
-            try{
-           var character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
-           if(character is null) throw new Exception($"Character with ID '{id}' not found.");
+            try
+            {
+                var character = await _context.Characters.Include(c => c.User).FirstOrDefaultAsync(c => c.Id == id && c.User!.Id == GetUserId());
+                if (character is null || character.User!.Id != GetUserId()) throw new Exception($"Character with ID '{id}' not found.");
 
-           _context.Characters.Remove(character);
-           await _context.SaveChangesAsync();
+                _context.Characters.Remove(character);
+                await _context.SaveChangesAsync();
 
-           serviceResponse.Data = await _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();;
-            }  
-            catch(Exception ex){
+                serviceResponse.Data = await _context.Characters.Where(c => c.User!.Id == GetUserId()).Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync(); ;
+            }
+            catch (Exception ex)
+            {
                 serviceResponse.Success = false;
                 serviceResponse.Message = ex.Message;
             }
-           
-           return serviceResponse;
+
+            return serviceResponse;
         }
     }
 }
